Add description as new DatasetType property
Similar to #1143, this PR adds "description" from the product definition to a new property in `DatasetType`. This will make it easier to access programatically.

I've also used this change to simplify the `to_dict` functionality, and ensured that this adds both of the new properties to the dict (e.g. `license` and `description`).