Refactor GridSpec class
1. Use pixel units to defined grid, not CRS units

Using tile size in CRS units + resolution leads to a possibility of fractional
pixel size. If tile size is 7m and pixel size is 3m, what should the size of a
tile be in pixels? Then there are problem of numeric precision. Defining grid in
pixels using integer sizes doesn't have this problem.

2. Negative tile sizes not allowed

Previously one could use negative tile size to reverse the order for grid
indexes. This is confusing and was not documented well.

Instead `fplix/flipy` parameters are introduced.

3. Adding more convenient construction options
   - From some tile geometry, index and shape
   - Direct support for slippy tiles regime

4. Adding geojson output for debugging and visualization