Remove tile_buffer option from tiles_from_geopolygon
It was not documented well and I could not find any code in datacube repo that
used this feature to understand what was expected from it. Tests were also very
basic and not enough to understand expected behaviour.

Effect of adding buffering was:

1. You get extra tiles that are "close" but do not overlap with the query
2. Output tiles themselves were padded out, and so won't be tiles anymore, as
   they would be overlapping. I could not verify if that was desired effect from
   looking at the code that might be using that feature, but I think that
   behaviour is confusing.

If you need tiles that are close to query region but do not quite overlap, you
can just buffer the query polygon before passing it in.

If you want padded out tiles with overlap, you can buffer output geoboxes that
come out.