Deprecate iterable interface from Geometry class
`.geoms` is a replacement, iteration over sub-geometries has been
deprecated in shapely and will be removed soon.

Geometry not being an iterable allows for more convenience overloads
to be developed. For example implementing `Affine * Geometry` with
`__rmul__` was not working because `Afffine * Iterable[Any]` was taking
precedence, but only when Geometry was a multi-geometry of 2 elements.