Removing __array_interface__
__array_interface__ is deprecated by shapely and will soon be removed,
it does not apply uniformly across geometry types and makes it harder
to create numpy arrays of geometry objects which is sometimes usefull
to have. It's not used anywhere as far as I can tell and was just added
for completeness.

It's mostly usefull for converting LineString to array of points, but
same can be achieved with `numpy.asarray(g.coords)`.